Asianboii's Extraordinary UI 2.x Versions
Updates of a lot of functionalities. Not a whole lot to say here, since the 2.x versions are not as original as the 1.x and not as professional and cool as the latest 3.x versions. Yet these versions are a great milestone for me to come up with some more cool ideas. Stay tuned for the newest versions!