Asianboii's Extraordinary UI 3.0 Version!
This version comes with brand-new design programmed with brand-new code, and it's the second newest version so far. This version provides a new designed keyboard layout (see the image files in the directory for more information), along with more cool functionalities, most of them are used with the Fn key (by me, not the fn key already on the keyboard). There's still one more version which will be even cooler, so stay tuned.